feature: provide access to flow state in timeout event handler
Since timeout uses :dispatch-later, then the event(s) will always fire,
even when the flow may have already halted successfully or failed.
So that the handler can decide what to do at the time of the timeout,
it is handed a delay to the flow state (when db-path is not specified).
With this it can decide if the flow has already completed or not and in
turn choose if it wants to halt the flow or not.
In the later case, it should do so by dispatching one of the existing
halt rules.

Issue #2